- cambiado defaultprofile de sitio /uploads/ -cambios de diseño

+++ src/Edukagames/UserBundle/Util/SaveEraseFile.php
@@ -30,13 +30,16 @@ class SaveEraseFile
     
     public static function eraseDir($dir) 
     {
-    	foreach(glob($dir . '/*') as $file) {
-        if(is_dir($file))
-            eraseDir($file);
-        else
-            unlink($file);
-	    }
-	    rmdir($dir);
+    	if(file_exists($dir)){
+    		foreach(glob($dir . '/*') as $file) {
+    			if(is_dir($file))
+    				eraseDir($file);
+    			else
+    				unlink($file);
+    		}
+    		rmdir($dir);
+    	}
+
     }
 
 }
